Add PulseSync (used in iob-eth)
The PulseSync macro is a clock bridge that is used when a signal is only asserted for one cycle (it is constant for long periods of time, except for one cycle where it transitions and then goes back to the original value and repeats). The standard 2 FF clock crossing is not adequate to this cases, as the slower clock might fail to sample the signal.