axil bus transactions issue solved
It is difficult to use tasks as they only update outputs when exiting.
A high number of arguments needed to be passed and calling the task
became more onerous than writing inline code.

Inline code was adopted and read and write templates added in
hardware/simulation/axil.vt. Tested in i2s-tdm.