feat(parameters): Add support for LIB module python parameters With these modifications, the `modules` lists from the `submodules['*_setup']['modules']` in the `*_setup.py` file, accept optional parameters along with lib modules. The format for passing parameters is a python tuple, where the second argument is the parameter: `('<module_name>',<parameter_obj>)` The `parameter_obj` may be any python object, including a list of parameters. The imported module can access the `parameter_obj` passed, via the variable named `module_parameters`.
For example, if we set the `modules` list as: `submodules['hw_setup']['modules']= [ ('iob_fifo_sync', (4,8)) ]`
The `iob_fifo_sync.py` python module will receive a variable named `module_parameters` with the value `(4,8)`.
(This value example is the tuple with two elements, 4 and 8, that we passed in the modules list).

If we want to manipulate the `hw_modules` variable of the `iob_fifo_sync.py` based on the `module_parameters` received, we could, for example, add the following code to that module:
```
if 'module_parameters' in globals():
	# If the first element of the tuple passed is less that 5 (example tuple '(4,8)' -> first element is '4')
	if module_parameters[0] < 5:
		hw_modules += ['iob_ram_2p_asym_wgtr']
	else:
		hw_modules += ['iob_ram_2p_asym_wler']
```