feat(linter): Add hack to `verilog-lint.sh` to make verible check macros. This commits adds a hack to the `verilog-lint.sh` script that forces verible to analyse Verilog macros.
It seems that Verible only analyses macros if there is a syntax error in
the Verilog sources. Therefore, the hack consists of adding a syntax
error to the end of the combined verilog file and running the linter.
If it detects any errors related to macros, it prints them.
If it does not detect any errors related to macros, it removes the
syntax error from the combined file and runs the linter again.