feat(linter): Remove verible; Add verilator linter. Update `verilog-lint.sh` to use the `verilator` linter, or optionally `svlint`.
Verible is no longer used because it has bugs parsing complex macros and
cannot include verilog headers.

The `verilog-lint.sh` script consists mostrly of python3 code, therefore
it may be replaced by a python script. For now, the bash code only sets
and executes the linter command.

The python script will receive a list of files, find out their
directories, generate a list of combinations for those directories, and
create a list of files for each of those combinations.
For each combination, it runs the linter on the corresponding file list.

The reason for the directory combinations is to prevent mixing files from
incompatible directories, such as the `fpga/src/` and `simulation/src/`.

The `svlint` is system verilog linter (Nix package available) that can
be used in this script instead (or along with) verilator.

For now, both of those linters throw many erros and warnings with the IOb-SoC system.
We should update the linter rules or fix those issues.