Correct bugs and add iob_demux2
Corrected a bug in submodule_utils.py and another one in iob_split.py.
The latter happened because 3 signals were not being distributed to all
the followers - only to the first one.

To correct that bug, demux2 was added, working the same way as iob_mux -
by storing the signals into a register instead of just assigning wires.